LayoutTests/media/video-src-expected.txt

EVENT(loadstart)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test') OK
EXPECTED (stripExtension(relativeURL(video.src)) == 'content/test') OK
EXPECTED (stripExtension(video.getAttribute('src')) == 'content/test') OK
END OF TEST
